# rad-app/src/main/resources/rad-app/com/apihug/api.yaml
openapi: 3.0.1
info:
  title: rad
  description: apihug rapid application development template
  contact:
    name: developer@apihug.com
    url: https://github.com/apihug/
    email: developer@apihug.com
  version: 0.1.1-SNAPSHOT
...
paths:
...
  /api/platforms/tenants/create:
    post:
      tags:
      - tenant
      summary: 创建租户
      description: 创建新租户
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "200":
          description: 创建新租户
          content:
            application/json:
              schema:
                type: object
                allOf:
                - $ref: "#/components/schemas/CommonResult"
                properties:
                  data:
                    $ref: "#/components/schemas/TenantSummary"
      x-hope-operation:
        requestClz: com.apihug.rad.api.platform.CreateTenantRequest
        responseClz: com.apihug.rad.api.platform.TenantSummary
        priority: HIGH
        authorization:
          rbac:
            predefinedRoleChecker: PLATFORM_MANAGER
            authorities:
            - TENANT_CREATE
            combinator: AND
  /api/platforms/tenants/get/{tenantId}:
    get:
      tags:
      - tenant
      summary: 获取租户详情
      description: 获取租户详情
      parameters:
      - name: tenantId
        in: path
        required: true
        allowEmptyValue: true
        schema:
          type: number
          format: int64
      responses:
        "200":
          description: 获取租户详情
          content:
            application/json:
              schema:
                type: object
                allOf:
                - $ref: "#/components/schemas/CommonResult"
                properties:
                  data:
                    $ref: "#/components/schemas/TenantDetail"
      x-hope-operation:
        responseClz: com.apihug.rad.api.platform.TenantDetail
        priority: LOW
        authorization:
          rbac:
            predefinedRoleChecker: PLATFORM
  /api/platforms/tenants/update/{tenantId}:
    put:
      tags:
      - tenant
      summary: 更新租户
      description: 更新租户信息
      parameters:
      - name: tenantId
        in: path
        required: true
        allowEmptyValue: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/UpdateTenantRequest"
      x-hope-operation:
        requestClz: com.apihug.rad.api.platform.UpdateTenantRequest
        priority: MIDDLE
        authorization:
          rbac:
            predefinedRoleChecker: PLATFORM_MANAGER
            authorities:
            - TENANT_UPDATE
            combinator: AND
  /api/platforms/tenants/{tenantId}/disable:
    delete:
      tags:
      - tenant
      summary: 停用租户
      description: 停用租户
      parameters:
      - name: tenantId
        in: path
        required: true
        allowEmptyValue: true
        schema:
          type: integer
          format: int32
      x-hope-operation:
        priority: HIGH
        authorization:
          rbac:
            predefinedRoleChecker: PLATFORM
            authorities:
            - TENANT_DISABLE
            combinator: AND
  /api/platforms/tenants/{tenantId}/configure:
    post:
      tags:
      - tenant
      summary: 配置租户
      description: 配置租户功能
      parameters:
      - name: tenantId
        in: path
        required: true
        allowEmptyValue: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/ConfigureTenantRequest"
      x-hope-operation:
        requestClz: com.apihug.rad.api.platform.ConfigureTenantRequest
        priority: MIDDLE
        authorization:
          rbac:
            predefinedRoleChecker: PLATFORM
            authorities:
            - TENANT_CONFIGURE
            combinator: AND
  /api/platforms/tenants/search:
    post:
      tags:
      - tenant
      summary: 搜索租户
      description: 搜索租户（分页）
      parameters:
      - name: pageRequest
        in: query
        $ref: "#/components/parameters/CommonPageRequest"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/SearchTenantsRequest"
      responses:
        "200":
          description: 搜索租户（分页）
          content:
            application/json:
              schema:
                type: object
                allOf:
                - $ref: "#/components/schemas/CommonResult"
                properties:
                  data:
                    type: object
                    allOf:
                    - $ref: "#/components/schemas/CommonPageableResult"
                    properties:
                      data:
                        type: array
                        description: "data list, payload of this pageable result"
                        items:
                          $ref: "#/components/schemas/TenantSummary"
      x-hope-operation:
        requestClz: com.apihug.rad.api.platform.SearchTenantsRequest
        responseClz: com.apihug.rad.api.platform.TenantSummary
        priority: LOW
        authorization:
          rbac:
            predefinedRoleChecker: PLATFORM
            authorities:
            - TENANT_CREATE
            combinator: AND
...
components:
  schemas:
...
    CommonPageableResult:
      type: object
      description: Pageable result wrapper
      properties:
        pageIndex:
          type: integer
          format: int32
          description: Current page index (0-based)
          example: 0
        pageSize:
          type: integer
          format: int32
          description: Number of items per page
          example: 10
        totalCount:
          type: integer
          format: int64
          description: Total number of items
          example: 125
        totalPage:
          type: integer
          format: int32
          description: Total number of pages
          example: 13
      title: (Common)Pageable Result Wrapper
    CommonResult:
      type: object
      description: Standard API response wrapper
      properties:
        code:
          type: integer
          format: int32
          description: "Response status code: 0 indicates success; non-zero values\
            \ indicate errors."
          example: 0
        message:
          type: string
          description: Human-readable message providing additional context about the
            result.
          example: Operation succeeded
        errors:
          type: array
          description: List of error details (omitted if empty)
          items:
            $ref: "#/components/schemas/CommonError"
      title: (Common)Result Wrapper
...
    TenantStatusEnum:
      type: string
      description: 租户状态枚举
      enum:
      - ACTIVE
      - DISABLED
      - EXPIRED
      x-hope-component:
        x-enum-extensions:
          ACTIVE:
            name: ACTIVE
            code: 1
            message: active
            message2: 活跃
          DISABLED:
            name: DISABLED
            code: 2
            message: disabled
            message2: 禁用
          EXPIRED:
            name: EXPIRED
            code: 3
            message: expired
            message2: 过期
        compile:
          protoFile: com/apihug/rad/infra/tenant/constant.proto
          name: TenantStatusEnum
          locationLine: 8
          locationColumn: 1
        isEnum: true
        isError: false
        clz: com.apihug.rad.infra.tenant.TenantStatusEnum
...
    SearchTenantsRequest:
      type: object
      description: 搜索租户请求
      properties:
        keyword:
          type: string
          description: 搜索关键词（租户名称或代码）
          x-hope-field:
            compile:
              locationLine: 408
              locationColumn: 3
              name: keyword
              type: java.lang.String
        status:
          type: object
          $ref: "#/components/schemas/TenantStatusEnum"
          description: 租户状态筛选
          enum:
          - ACTIVE
          - DISABLED
          - EXPIRED
          x-hope-field:
            compile:
              locationLine: 413
              locationColumn: 3
              name: status
              type: com.apihug.rad.infra.tenant.TenantStatusEnum
      title: SearchTenantsRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/platform/tenant.proto
          name: SearchTenantsRequest
          locationLine: 400
          locationColumn: 1
        clz: com.apihug.rad.api.platform.SearchTenantsRequest
...
    ConfigureTenantRequest:
      type: object
      description: 配置租户请求
      properties:
        maxMembers:
          type: integer
          format: int32
          description: 最大成员数
          minimum: 1.0
          x-hope-field:
            compile:
              locationLine: 266
              locationColumn: 3
              name: max_members
              type: java.lang.Integer
        maxStorageMb:
          type: number
          format: int64
          description: 最大存储空间（MB）
          minimum: 1.0
          x-hope-field:
            compile:
              locationLine: 272
              locationColumn: 3
              name: max_storage_mb
              type: java.lang.Long
        enabledModules:
          type: array
          description: 启用的功能模块
          items:
            type: string
          x-hope-field:
            compile:
              locationLine: 278
              locationColumn: 3
              name: enabled_modules
              type: java.lang.String
        expiryDate:
          type: string
          format: date-time
          description: 到期时间
          x-hope-field:
            dateFormat: ISO_LOCAL_DATE_TIME
            compile:
              locationLine: 283
              locationColumn: 3
              name: expiry_date
              type: java.time.LocalDateTime
      title: ConfigureTenantRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/platform/tenant.proto
          name: ConfigureTenantRequest
          locationLine: 258
          locationColumn: 1
        clz: com.apihug.rad.api.platform.ConfigureTenantRequest
...
    UpdateTenantRequest:
      type: object
      description: 更新租户请求
      properties:
        tenantName:
          type: string
          description: 租户名称
          maxLength: 100
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 228
              locationColumn: 3
              name: tenant_name
              type: java.lang.String
        contactEmail:
          type: string
          description: 联系人邮箱
          maxLength: 100
          minLength: 1
          x-hope-field:
            mock:
              nature: EMAIL
            compile:
              locationLine: 236
              locationColumn: 3
              name: contact_email
              type: java.lang.String
        contactPhone:
          type: string
          description: 联系人电话
          maxLength: 20
          x-hope-field:
            mock:
              nature: PHONE
            compile:
              locationLine: 244
              locationColumn: 3
              name: contact_phone
              type: java.lang.String
        status:
          type: object
          $ref: "#/components/schemas/TenantStatusEnum"
          description: 租户状态
          enum:
          - ACTIVE
          - DISABLED
          - EXPIRED
          x-hope-field:
            compile:
              locationLine: 251
              locationColumn: 3
              name: status
              type: com.apihug.rad.infra.tenant.TenantStatusEnum
      title: UpdateTenantRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/platform/tenant.proto
          name: UpdateTenantRequest
          locationLine: 220
          locationColumn: 1
        clz: com.apihug.rad.api.platform.UpdateTenantRequest
...
    TenantSummary:
      type: object
      description: 租户摘要信息
      properties:
        id:
          type: number
          format: int64
          description: 租户 ID
          minimum: 1.0
          x-hope-field:
            compile:
              locationLine: 298
              locationColumn: 3
              name: id
              type: java.lang.Long
        tenantCode:
          type: string
          description: 租户代码
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 305
              locationColumn: 3
              name: tenant_code
              type: java.lang.String
        tenantName:
          type: string
          description: 租户名称
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 312
              locationColumn: 3
              name: tenant_name
              type: java.lang.String
        status:
          type: object
          $ref: "#/components/schemas/TenantStatusEnum"
          description: 租户状态
          enum:
          - ACTIVE
          - DISABLED
          - EXPIRED
          x-hope-field:
            compile:
              locationLine: 319
              locationColumn: 3
              name: status
              type: com.apihug.rad.infra.tenant.TenantStatusEnum
      title: TenantSummary
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/platform/tenant.proto
          name: TenantSummary
          locationLine: 290
          locationColumn: 1
        clz: com.apihug.rad.api.platform.TenantSummary
...
    CreateTenantRequest:
      type: object
      description: 创建租户请求
      properties:
        tenantCode:
          type: string
          description: 租户代码
          maxLength: 50
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 171
              locationColumn: 3
              name: tenant_code
              type: java.lang.String
        tenantName:
          type: string
          description: 租户名称
          maxLength: 100
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 179
              locationColumn: 3
              name: tenant_name
              type: java.lang.String
        contactEmail:
          type: string
          description: 联系人邮箱
          maxLength: 100
          minLength: 1
          x-hope-field:
            mock:
              nature: EMAIL
            compile:
              locationLine: 187
              locationColumn: 3
              name: contact_email
              type: java.lang.String
        contactPhone:
          type: string
          description: 联系人电话
          maxLength: 20
          x-hope-field:
            mock:
              nature: PHONE
            compile:
              locationLine: 195
              locationColumn: 3
              name: contact_phone
              type: java.lang.String
        status:
          type: object
          $ref: "#/components/schemas/TenantStatusEnum"
          description: 租户状态
          enum:
          - ACTIVE
          - DISABLED
          - EXPIRED
          x-hope-field:
            compile:
              locationLine: 202
              locationColumn: 3
              name: status
              type: com.apihug.rad.infra.tenant.TenantStatusEnum
        isPlatform:
          type: boolean
          description: 是否为平台租户
          x-hope-field:
            compile:
              locationLine: 208
              locationColumn: 3
              name: is_platform
              type: java.lang.Boolean
        description:
          type: string
          description: 租户描述
          maxLength: 500
          x-hope-field:
            compile:
              locationLine: 213
              locationColumn: 3
              name: description
              type: java.lang.String
      title: CreateTenantRequest
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/platform/tenant.proto
          name: CreateTenantRequest
          locationLine: 163
          locationColumn: 1
        clz: com.apihug.rad.api.platform.CreateTenantRequest
...
    TenantDetail:
      type: object
      description: 租户详情信息
      properties:
        id:
          type: number
          format: int64
          description: 租户 ID
          minimum: 1.0
          x-hope-field:
            compile:
              locationLine: 334
              locationColumn: 3
              name: id
              type: java.lang.Long
        tenantCode:
          type: string
          description: 租户代码
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 341
              locationColumn: 3
              name: tenant_code
              type: java.lang.String
        tenantName:
          type: string
          description: 租户名称
          minLength: 1
          x-hope-field:
            mock:
              nature: NAME
            compile:
              locationLine: 348
              locationColumn: 3
              name: tenant_name
              type: java.lang.String
        contactEmail:
          type: string
          description: 联系人邮箱
          x-hope-field:
            mock:
              nature: EMAIL
            compile:
              locationLine: 355
              locationColumn: 3
              name: contact_email
              type: java.lang.String
        contactPhone:
          type: string
          description: 联系人电话
          x-hope-field:
            mock:
              nature: PHONE
            compile:
              locationLine: 361
              locationColumn: 3
              name: contact_phone
              type: java.lang.String
        status:
          type: object
          $ref: "#/components/schemas/TenantStatusEnum"
          description: 租户状态
          enum:
          - ACTIVE
          - DISABLED
          - EXPIRED
          x-hope-field:
            compile:
              locationLine: 367
              locationColumn: 3
              name: status
              type: com.apihug.rad.infra.tenant.TenantStatusEnum
        maxMembers:
          type: integer
          format: int32
          description: 最大成员数
          x-hope-field:
            compile:
              locationLine: 373
              locationColumn: 3
              name: max_members
              type: java.lang.Integer
        maxStorageMb:
          type: number
          format: int64
          description: 最大存储空间（MB）
          x-hope-field:
            compile:
              locationLine: 378
              locationColumn: 3
              name: max_storage_mb
              type: java.lang.Long
        isPlatform:
          type: boolean
          description: 是否为平台租户
          x-hope-field:
            compile:
              locationLine: 383
              locationColumn: 3
              name: is_platform
              type: java.lang.Boolean
        description:
          type: string
          description: 租户描述
          x-hope-field:
            compile:
              locationLine: 388
              locationColumn: 3
              name: description
              type: java.lang.String
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          x-hope-field:
            dateFormat: ISO_LOCAL_DATE_TIME
            compile:
              locationLine: 393
              locationColumn: 3
              name: created_at
              type: java.time.LocalDateTime
      title: TenantDetail
      x-hope-component:
        compile:
          protoFile: com/apihug/rad/api/platform/tenant.proto
          name: TenantDetail
          locationLine: 326
          locationColumn: 1
        clz: com.apihug.rad.api.platform.TenantDetail
...
  parameters:
    CommonPageRequest:
      name: pageRequest
      in: query
      description: Page Request parameter
      schema:
        type: object
        description: Pagination parameters for querying a collection of resources.
        properties:
          page:
            type: integer
            format: int32
            description: "Zero-based page index (e.g., 0 for the first page). Must\
              \ be >= 0."
            minimum: 0
            title: Page Index
          size:
            type: integer
            format: int32
            default: 20
            description: Number of items per page. Defaults to 20 if unspecified.
            maximum: 1024
            minimum: 1
            title: Page Size
          sort:
            type: array
            description: |-
              Sorting instructions in the format: `property,(asc|desc)`.
              Multiple sort criteria may be provided (e.g., `["name,asc", "createdAt,desc"]`).
              If direction is omitted, ascending (`asc`) is assumed by default.
            items:
              type: string
        title: (Common)Page Request Parameters
...